MVP-6451-patch: Refactor Auth configuration & document
