+++ packages/notifi-frontend-client/lib/client/index.ts
@@ -1,10 +1,6 @@
 export * from './NotifiFrontendClient';
 export * from './NotifiSmartLinkClient';
-export {
-  instantiateFrontendClient,
-  newSmartLinkClient,
-  UserParams,
-} from './clientFactory';
+export { instantiateFrontendClient, newSmartLinkClient } from './clientFactory';
 export * from './deprecated';
 // export * from './AuthManager';
 export * from './auth';
